fix: move received timestamp closer to when frame is acquired
The frame timestamp used is not the timestamp at the moment
of the frame capture.This might be an issue when synchronizing
with the transforms.
At the moment the frame timestamp is generated on the syskit side,
when the base::frame is created. This PR to moves the timestamp
acquisition as close as possible to when we receiving the frames.